refactor(tools): update path for get annotation
